Changed routes to make Login the default
